+++ packages/shared_preferences/shared_preferences/example/android/app/src/main/java/io/flutter/plugins/sharedpreferencesexample/EmbeddingV1Activity.java
@@ -5,14 +5,18 @@
 package io.flutter.plugins.sharedpreferencesexample;
 
 import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import dev.flutter.plugins.integration_test.IntegrationTestPlugin;
+import io.flutter.plugins.sharedpreferences.SharedPreferencesPlugin;
 
-public class EmbeddingV1Activity extends FlutterActivity {
+@SuppressWarnings("deprecation")
+public class EmbeddingV1Activity extends io.flutter.app.FlutterActivity {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+    IntegrationTestPlugin.registerWith(
+        registrarFor("dev.flutter.plugins.integration_test.IntegrationTestPlugin"));
+    SharedPreferencesPlugin.registerWith(
+        registrarFor("io.flutter.plugins.sharedpreferences.SharedPreferencesPlugin"));
   }
 }