+++ packages/video_player/video_player/test/video_player_initialization_test.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.8
+
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:video_player/video_player.dart';
@@ -19,6 +21,6 @@ void main() {
       'https://127.0.0.1',
     );
     await controller.initialize();
-    expect(fakeVideoPlayerPlatform.calls.first.method, 'init');
+    expect(fakeVideoPlayerPlatform.calls.first, 'init');
   });
 }