+++ packages/share/CHANGELOG.md
@@ -1,3 +1,36 @@
+## 0.6.5+2
+
+* Keep handling deprecated Android v1 classes for backward compatibility.
+
+## 0.6.5+1
+
+* Avoiding uses unchecked or unsafe Object Type Casting
+
+## 0.6.5
+
+* Added support for sharing files
+
+## 0.6.4+5
+
+* Update package:e2e -> package:integration_test
+
+## 0.6.4+4
+
+* Update package:e2e reference to use the local version in the flutter/plugins
+  repository.
+
+## 0.6.4+3
+
+* Post-v2 Android embedding cleanup.
+
+## 0.6.4+2
+
+* Update lower bound of dart dependency to 2.1.0.
+
+## 0.6.4+1
+
+* Declare API stability and compatibility with `1.0.0` (more details at: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0).
+
 ## 0.6.4
 
 * Remove Android dependencies fallback.