+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,6 +1,31 @@
-## 1.0.6
+## 2.3.0
 
 * Add caching functionality to videos from network sources.
+
+## 2.2.0
+
+* Added option to set the video playback speed on the video controller.
+
+## 2.1.1
+
+* Fix mixWithOthers test channel.
+
+## 2.1.0
+
+* Add VideoPlayerOptions with audo mix mode
+
+## 2.0.2
+
+* Migrated tests to use pigeon correctly.
+
+## 2.0.1
+
+* Updated minimum Dart version.
+* Added class to help testing Pigeon communication.
+
+## 2.0.0
+
+* Migrated to [pigeon](https://pub.dev/packages/pigeon).
 
 ## 1.0.5